changing comment and explaining the mass increase in code over the past 24 hours, as had to change between VS code and spyder to aid in processing time
When going through my assesment last night I found it really difficult to process the code on pytest through the terminal, it ended up taking upwards of 30 minutes to do so per process.

This lead me to change part way through to VS code. This ment that when I switched it back over to Spyder this morning, the code had alot of changes which I didn't git commit individually, as I wasn't sure how without causing repo issues.

i used vs code to reduce the pytest processing time, particularly on Spyder and using the anaconda powershell terminal.

This was under the instruction of gemini and github, as it suggested a vertual processing space such as VS-Code and google Colab have a quicker processing time and have better descriptions of error messages/ linting.

Due to using vs code for some of the formative (as i had a file issue over easter) i was somewhat familar with it, leading it to be by choice to switch back and fourth from Spyder / Vs code to complete the assessment the best i can.

so far i have passed 5 tests and failed 2, so will work out how to fix those 2 failing tests and then lint my work/ make it neater