ADDING KEY NOTE FOR RUNNING CODE
As of right now the code is passing on MAC OS and Lineux opperating systems - when tested in the university library.

When doing the command python -m pytest test/test_terrain.py in a windows opperating system , the processing time takes upwards of half a hour - currently passing five tests but failing two inconsistantly.

However, when testing on a MacOS and Lineux operating system all tests pass constantly, taking just over 7 minutes to process the code.

I would prefer if the exam can be ran in eaither Lineux or Mac OS if possible.

I read the inconsistancy in pytest in windows is due to rasterio and fiona ( Key functions i imported into my code to get the raster and shape files to work ) having a conflicted system. leading pytest to be less effective , also leading to test_regression and test_verbose to fail at points.

Thankyou for your consideration when running this exam