Fixing a linting issue
Due to the tempory change in the function def caculate_slope to def calculate_slope_vectorized, i needed to change the variable name in my def main() , specifically on line 306.

I decided to attach a note on it to allow myself to recognise the change when I read back through my code, partically if the vectorized version of the variable fails and i would need to change it back.