linting error which i was correcting
linting error flsgged the misnig empty line space below line 371. This has been flagged as a linting issue as there is nothing for the code to output onto a the very end of the file.

Gemini's recomdeded fix for this error was adding a line of whitespace - therefore creating the blank line 372.

This hopefully should solve one of my linting issues and get me one step closer to passing the linter test.

right now my linter score is 9.367 /10. To hopefully pass the required linting test the score needs to be 10/10 , making me 0.533 points away currently.